[refactor] разделить View и ViewModels
