+++ iOS-FakeNFT-Extended/UI/StatisticsFlow/Statistics/ViewModels/StatisticsViewModel.swift
@@ -1,0 +1,53 @@
+import SwiftUI
+
+@MainActor
+@Observable
+final class StatisticsViewModel {
+    enum SortOrder: String, CaseIterable, Identifiable {
+        var id: String { rawValue }
+        case name = "По имени"
+        case rating = "По рейтингу"
+    }
+    enum State: Equatable {
+        case `default`
+        case loading
+        case loaded
+        case error
+    }
+    var allUsers: [UserViewModel] = []
+    var sortOrder: SortOrder = .rating
+    var showSortMenu: Bool = false
+    var state: State = .default
+    var loadUsers: () async throws -> [UserViewModel] = { [] }
+}
+
+extension StatisticsViewModel {
+    var hasError: Bool {
+        get { state == .error }
+        set { state = .default }
+    }
+    func load() async {
+        state = .loading
+        do {
+            allUsers = try await loadUsers()
+            state = .loaded
+        } catch {
+            state = .error
+            print(error.localizedDescription)
+        }
+    }
+    var sortedUsers: [UserViewModel] {
+        let sorted = allUsers.sorted { (user1, user2) -> Bool in
+            switch sortOrder {
+            case .name:
+                user1.name < user2.name
+            case .rating:
+                user1.itemCount > user2.itemCount
+            }
+        }
+        for (index, user) in sorted.enumerated() {
+            user.position = index + 1
+        }
+        return Array(sorted)
+    }
+}
